Tests/LibWeb/Text/expected/css/scrollable-overflow-recalc-pseudo-element.txt

initial scrollHeight: 100
scrollHeight after pseudo-element transform change: 550
relayouts performed: 0
